add disable replay config
